src/driver/pyproject.toml

[build-system]
requires = ["uv_build>=0.8.20,<0.9.0"]
build-backend = "uv_build"

[project]
name = "alpasim_driver"
version = "0.7.0"
description = "Adapters for open-source drivers in Alpasim"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "vam",
  "alpamayo_r1",
  "einops",
  "numpy",
  "opencv-python",
  "torch",
  "pillow",
  "rich",
  "click",
  "grpcio",
  "hydra-core",
  "omegaconf",
  "alpasim-grpc",
  "beartype",
  "jaxtyping",
  "timm",
]

[project.scripts]
alpasim_driver_main = "alpasim_driver.main:main"

[tool.uv.sources]
vam = { git = "https://github.com/valeoai/VideoActionModel.git" }
alpamayo_r1 = { git = "https://github.com/NVlabs/alpamayo.git", rev = "74750b0bd86605ced9faed9fe83434b07f95e9dc"}
alpasim-grpc = { workspace = true }
